+++ app/src/main/java/zapsolutions/zap/LandingActivity.java
@@ -1,15 +1,13 @@
 package zapsolutions.zap;
 
 import android.app.AlertDialog;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 
 import zapsolutions.zap.baseClasses.App;
 import zapsolutions.zap.baseClasses.BaseAppCompatActivity;
-import zapsolutions.zap.connection.manageWalletConfigs.Cryptography;
-import zapsolutions.zap.pin.PinEntryActivity;
+import zapsolutions.zap.util.PinScreenUtil;
 import zapsolutions.zap.util.PrefsUtil;
 import zapsolutions.zap.util.RefConstants;
 import zapsolutions.zap.util.ZapLog;
@@ -69,43 +67,16 @@ public class LandingActivity extends BaseAppCompatActivity {
         PrefsUtil.edit().putInt(PrefsUtil.SETTINGS_VERSION, RefConstants.CURRENT_SETTINGS_VERSION).commit();
 
         if (PrefsUtil.isWalletSetup()) {
+            PinScreenUtil.askForAccess(this, () -> {
+                Intent homeIntent = new Intent(this, HomeActivity.class);
+                homeIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
+                startActivity(homeIntent);
+            });
 
-            if (PrefsUtil.isPinEnabled()){
-                // Go to PIN entry screen
-                Intent pinIntent = new Intent(this, PinEntryActivity.class);
-                pinIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
-                startActivity(pinIntent);
-            } else {
-
-                // Check if pin is active according to key store
-                boolean isPinActive = false;
-                try {
-                    isPinActive =  new Cryptography(this).isPinActive();
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-
-                // Only allow access if pin is not active in key store!
-                if (isPinActive) {
-                    // According to the key store, the pin is still active. This happens if the pin got deleted from the prefs without also removing the keystore entry.
-                    new AlertDialog.Builder(this)
-                            .setMessage(R.string.error_pin_deactivation_attempt)
-                            .setCancelable(false)
-                            .setPositiveButton(R.string.continue_string, new DialogInterface.OnClickListener() {
-                                public void onClick(DialogInterface dialog, int whichButton) {
-                                    LandingActivity.this.finish();
-                                }
-                            }).show();
-                } else {
-                    Intent homeIntent = new Intent(this, HomeActivity.class);
-                    homeIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_NEW_TASK);
-                    startActivity(homeIntent);
-                }
-            }
         } else {
 
             // Clear connection data if something is there
-            PrefsUtil.edit().putString(PrefsUtil.WALLET_CONFIGS, "").commit();
+            PrefsUtil.edit().remove(PrefsUtil.WALLET_CONFIGS).commit();
 
 
             Intent homeIntent = new Intent(this, HomeActivity.class);
